# Reimplement AutoDiff from the ground up