--- a/common/ad/BUILD.bazel
+++ b/common/ad/BUILD.bazel
@@ -22,6 +22,7 @@ drake_cc_library(
         "internal/derivatives_xpr.cc",
         "internal/partials.cc",
         "internal/standard_operations.cc",
+        "internal/static_unit_vector.cc",
     ],
     hdrs = [
         "auto_diff.h",
@@ -30,9 +31,11 @@ drake_cc_library(
         "internal/eigen_specializations.h",
         "internal/partials.h",
         "internal/standard_operations.h",
+        "internal/static_unit_vector.h",
     ],
     deps = [
         "//common:essential",
+        "//common:reset_after_move",
     ],
 )
 
@@ -48,6 +51,14 @@ drake_cc_googletest(
     deps = [
         ":auto_diff",
         "//common/test_utilities:eigen_matrix_compare",
+    ],
+)
+
+drake_cc_googletest(
+    name = "eigen_specializations_heap_test",
+    deps = [
+        ":auto_diff",
+        "//common/test_utilities:limit_malloc",
     ],
 )
 
@@ -72,8 +83,16 @@ drake_cc_googletest(
 drake_cc_googletest(
     name = "partials_heap_test",
     deps = [
-        "//common/ad:auto_diff",
+        ":auto_diff",
         "//common/test_utilities:limit_malloc",
+    ],
+)
+
+drake_cc_googletest(
+    name = "static_unit_vector_test",
+    deps = [
+        ":auto_diff",
+        "//common/test_utilities:eigen_matrix_compare",
     ],
 )
 
@@ -133,6 +152,14 @@ drake_cc_googletest(
     deps = [
         ":auto_diff",
         ":standard_operations_test_h",
+    ],
+)
+
+drake_cc_googletest(
+    name = "standard_operations_heap_test",
+    deps = [
+        ":auto_diff",
+        "//common/test_utilities:limit_malloc",
     ],
 )