--- a/common/ad/internal/partials.h
+++ b/common/ad/internal/partials.h
@@ -1,12 +1,111 @@
 #pragma once
 
+#include <atomic>
+#include <cstdlib>
+#include <utility>
+
 #include "drake/common/ad/internal/derivatives_xpr.h"
+#include "drake/common/drake_assert.h"
 #include "drake/common/drake_copyable.h"
 #include "drake/common/eigen_types.h"
+#include "drake/common/reset_after_move.h"
 
 namespace drake {
 namespace ad {
 namespace internal {
+
+/* Reference-counted heap storage for an array of doubles, for use by the
+Partials class later in this file. The storage can be empty (null). */
+class CowVec {
+ public:
+  /* Creates empty (null) storage. */
+  CowVec() = default;
+
+  /* Allocates new storage of the given size, but does not initialize it.
+  If the size is zero, the storage will be empty (null). */
+  static CowVec Allocate(int size);
+
+  /* Steals the storage from `other`. */
+  CowVec(CowVec&& other) noexcept {
+    data_ = other.data_;
+    other.data_ = nullptr;
+  }
+
+  /* Steals the storage from `other`. */
+  CowVec& operator=(CowVec&& other) noexcept {
+    if (this != &other) {
+      if (data_ != nullptr) {
+        DecrementUseCount();
+      }
+      data_ = other.data_;
+      other.data_ = nullptr;
+    }
+    return *this;
+  }
+
+  /* Borrows the storage from `other` (incrementing use_count). */
+  CowVec(const CowVec& other) noexcept { data_ = other.borrow_data(); }
+
+  /* Borrows the storage from `other` (incrementing use_count). */
+  CowVec& operator=(const CowVec& other) noexcept {
+    if (this != &other) {
+      if (data_ != nullptr) {
+        DecrementUseCount();
+      }
+      data_ = other.borrow_data();
+    }
+    return *this;
+  }
+
+  /* Decrements the storage use_count, freeing iff it reached zero. */
+  ~CowVec() {
+    if (data_ != nullptr) {
+      DecrementUseCount();
+    }
+  }
+
+  /* Resets this back to null.
+  @pre data() != nullptr */
+  void ResetAssumingNonNull() {
+    DRAKE_ASSERT(data_ != nullptr);
+    DecrementUseCount();
+    data_ = nullptr;
+  }
+
+  /* Returns the double array storage (or null, when empty). */
+  const double* data() const { return data_; }
+  double* mutable_data() { return data_; }
+
+  /* Returns the current use count. Be careful when using this for control
+  flow. Values other than 0 or 1 are subject to TOCTOU races. */
+  int64_t get_use_count() const {
+    return (data_ == nullptr) ? 0 : mutable_use_count().load();
+  }
+
+ private:
+  using UseCount = std::atomic_int_fast64_t;
+
+  /* Returns a mutable reference to our atomic use_count integer.
+  @pre data_ is non-null (the use_count is part of the managed block). */
+  UseCount& mutable_use_count() const {
+    DRAKE_ASSERT(data_ != nullptr);
+    return *reinterpret_cast<UseCount*>(reinterpret_cast<char*>(data_) -
+                                        sizeof(UseCount));
+  }
+
+  /* For use by our copy constructor and copy assignment.
+  Increments the use_count (when present) and returns the data pointer. */
+  double* borrow_data() const {
+    if (data_ != nullptr) {
+      ++mutable_use_count();
+    }
+    return data_;
+  }
+
+  void DecrementUseCount();
+
+  double* data_{nullptr};
+};
 
 /* A vector of partial derivatives, optimized for use with Drake's AutoDiff.
 
@@ -51,10 +150,21 @@ class Partials {
   /* Constructs a vector with a copy of the given value. */
   explicit Partials(const Eigen::Ref<const Eigen::VectorXd>& value);
 
+  /* Constructs a vector with a copy of the given value.
+  @pre value.size() >= 2 */
+  explicit Partials(const ad::DerivativesConstXpr& value);
+
   ~Partials() = default;
 
   /* Returns the size of this vector. */
-  int size() const { return storage_.size(); }
+  int size() const {
+    // This is a careful representational choice so that calculating the
+    // effective size() from magic_size_ is just two branchless instructions.
+    // In particular, it compiles down to `(magic_size_ >> 31) ^ magic_size_`.
+    // The other obvious alternative would be to use `std::abs(magic_size_)`
+    // but that adds one extra instruction to tweak the complement by 1.
+    return (magic_size_ >= 0) ? int{magic_size_} : (-magic_size_ - 1);
+  }
 
   /* Updates `this` to be the same size as `other`.
   If `this` and `other` are already the same size then does nothing.
@@ -64,8 +174,30 @@ class Partials {
   Otherwise, throws an exception for mismatched sizes. */
   void MatchSizeOf(const Partials& other);
 
+  /* Returns true when this denotes a value known to be zero vector -- either a
+  zero-sized vector, or a vector filled with zeros. This might return false in
+  the rare case of a dense vector addition that canceled out to zero. */
+  bool is_known_zero() const { return magic_size_ <= 0; }
+
+  /* Returns an Eigen-compatible view into this vector. */
+  ad::DerivativesConstXpr make_const_xpr() const;
+
+  /* Returns an Eigen-compatible mutable view into this vector, including
+  resizing. This is expensive (makes a copy). */
+  ad::DerivativesMutableXpr MakeMutableXpr();
+
   /* Set this to zero. */
-  void SetZero();
+  void SetZero() {
+    if (magic_size_ <= 0) {
+      return;
+    }
+    magic_size_ = -magic_size_ - 1;
+    if (unit_ == 0) {
+      // Freeing the storage immediately is important for performance so that
+      // future copies of `this` will be cheap (no operations on use_count()).
+      storage_.ResetAssumingNonNull();
+    }
+  }
 
   /* Scales this vector by the given amount. */
   void Mul(double factor);
@@ -74,24 +206,55 @@ class Partials {
   void Div(double factor);
 
   /* Adds `other` into `this`. */
-  void Add(const Partials& other);
+  void Add(const Partials& other) { AddScaled(1.0, other); }
 
   /* Adds `scale * other` into `this`. */
-  void AddScaled(double scale, const Partials& other);
-
-  /* Returns an Eigen-compatible view into this vector. */
-  ad::DerivativesConstXpr make_const_xpr() const;
-
-  /* Returns the underlying storage vector (mutable). This may involve O(size)
-  multiplications. */
-  Eigen::VectorXd& GetRawStorageMutable();
+  void AddScaled(double scale, const Partials& other) {
+    // XXX This logic needs better comments.
+    // XXX Should we be checking for matched sizes in Debug builds?
+    if (other.magic_size_ <= 0) {
+      if (magic_size_ == 0) {
+        magic_size_ = other.magic_size_;
+      }
+      return;
+    }
+    if (scale == 0.0) {
+      return;
+    }
+    if (is_known_zero()) {
+      // Borrow from `other` with no new allocations.
+      *this = other;
+      // N.B. Only _after_ copying can we apply the scale.
+      Mul(scale);
+      return;
+    }
+    AddScaledImpl(scale, other);
+  }
 
  private:
+  /* The non-inline implementation of AddScaled(). */
+  void AddScaledImpl(double, const Partials&);
+
+  /* Returns true iff this represents a (scaled) unit vector. */
+  bool is_unit() const { return unit_ > 0; }
+
+  /* Returns the index of the non-zero element of this (scaled) unit vector.
+  @pre is_unit() */
+  int get_unit_index() const {
+    DRAKE_ASSERT(unit_ > 0);
+    return unit_ - 1;
+  }
+
   void ThrowIfDifferentSize(const Partials& other);
 
-  // Our effective value is `coeff_ * storage_`; we store them separately so
-  // that re-scaling is fast (we can just scale the coeff).
-  //
+  // Our MutableXpr type is allowed to set us via a backreference.
+  friend ad::DerivativesMutableXpr;
+  ad::DerivativesMutableXpr SetFrom(
+      const Eigen::Ref<const Eigen::VectorXd>& other);
+
+  void CheckInvariants() const;
+
+  // XXX re-home this comment.
   // We maintain an invariant that `coeff_` is always finite. If a modification
   // to it (e.g., multiplication by a factor) would cause it to become non-
   // finite, then instead of multiplying the factor into `coeff_`, instead we
@@ -99,8 +262,12 @@ class Partials {
   // through the non-zero terms in `storage_`). This is required to meet our API
   // contract of "any zero values will remain zero, even if the factor is ±∞ or
   // NaN" per our class overview.
+
+  // Refer to CheckInvariants() commentary in the cc file for what these denote.
+  reset_after_move<int32_t> magic_size_{0};
+  int32_t unit_{0};
   double coeff_{0.0};
-  Eigen::VectorXd storage_;
+  CowVec storage_;
 };
 
 }  // namespace internal