… common/ad/_zzz_notes.txt
… common/ad/internal/partials.cc
… common/ad/test/eigen_specializations_heap_test.cc
… common/ad/test/partials_heap_test.cc
… common/ad/test/partials_test.cc
… common/ad/test/standard_operations_heap_test.cc