Relates to #16877 and #16877.