@jwnimmer-tri 2022-07-04
This is a rough cut for now, with some missing pieces.  There are also a few bugs still showing up in tests, which indicates some of the computations might be wrong -- so possibly this is faster in some places by cheating (giving wrong answers) rather than for a good reason.  However, a lot of unit tests _do_ pass.  Even with those caveats, I think the prototype shows promise and we should continue working.

---

Here's the latest Cassie results for mass matix, inverse dynamics, and forward dynamics:

![AutoDiff rewrite (time per call)](https://user-images.githubusercontent.com/17596505/177077606-6c32eee7-0011-4e07-90f4-ede254615826.svg)

The chart shows that not only is the throughput improved ~2x, the stddevs are narrower.

---

Here's the raw numbers, for now:

Case | Min Time | Relative reduction
-- | -- | --
MM master | 1854 us |  
MM PR | 883 us | 52.3%
ID master | 2793 us |  
ID PR | 1001 us | 64.2%
FD master | 4241 us |  
FD PR | 2478 us | 41.6%

---

It won't be part of this first prototype, but having our own AutoDiff scalar also paves the way to sharing partials storage across all of the elements in a `MatrixX<AutoDiff>` down the road.