@jwnimmer-tri 2022-07-04
Here's the quick sample from `/multibody/inverse_kinematics:position_constraint_experiment` for `IiwaPositionConstraintFixture`:

Benchmark | Time | Relative reduction
-|-|-
EvalAutoDiffMbpDouble_mean (master)   | 80 us   |
EvalAutoDiffMbpDouble_mean (PR)       | 66 us   | 17%
EvalAutoDiffMbpAutoDiff_mean (master) | 2288 ns |
EvalAutoDiffMbpAutoDiff_mean (PR)     | 809 us  | 64%