@jwnimmer-tri 2022-07-06
The functional tests are passing now, though apparently there are still some memory leaks to chase down.

Here's the updated perf summary for the `//examples/multibody/cassie_benchmark:experiment` and `//multibody/inverse_kinematics:position_constraint_experiment` cases.

![AutoDiff rewrite (time per call)](https://user-images.githubusercontent.com/17596505/177440551-dffae80d-d70b-425e-bf96-bb8fa71effd9.svg)

Case | Cassie MM | Cassie ID | Cassie FD | IK MbpD | IK MbpAD
-- | -- | -- | -- | -- | --
Master (min) | 1854 us | 2793 us | 4241 us | 79 us | 2254 us
PR (min) | 815 us | 833 us | 2373 us | 68 us | 658 us
Reduction | 56.0% | 70.2% | 44.0% | 13.5% | 70.8%