@jwnimmer-tri 2022-07-11
Yes, I also have the same question myself.  (Another separate item to measure being the Eigen binop specializations.)

I belatedly realized that the benchmark coverage was not good enough, so I'm going to wait for #17539 and then start to re-measure.  I'm not sure how the overall speedup will come out, and the relative merits of each technique will probably change as well.

I was also thinking of merging this in stages, and e.g. keeping the CowVector at the tail end (using just a plain VectorXd in the meantime).  That would allow each stage to be separately measured for its speedup.

For each technique, I already did A/B testing before incorporating them, and all of them were relevant to _some_ degree (at least, with the old benchmark coverage).  I know that the separate scale factor and the sparse representation are critically important, and the move-optimized binops are moderately important.  I think CowVector will be important but I'm not 100% sure.