@jwnimmer-tri 2022-07-11
I am also curious to see what happens!  There is likely still going to be room for RotationMatrix-specific optimizations, though their complexity vs speedup benefit may weigh differently.

I think my first bid beyond this PR would be (2b) noted above.  If we kept the derivatives storage in a single 9N-sized block instead of 9 N-sized blocks, it's possible that we could use SIMD for the math (kind of like you did for the T=double case), where even large blocks of zeros are still pretty fast.

Something we could probably add to the CowVector representation is a (limited) inventory of non-zero partials, in case trading branches for flops is a win (e.g., for N>>10 and high sparsity).