@sherm1 2022-07-13
> If we kept the derivatives storage in a single 9N-sized block instead of 9 N-sized blocks, it's possible that we could use SIMD for the math

Actually this is already possible in the DifferentialMatrix form when two 3x3 column-ordered matrices are multiplied. That's one of the AVX functions we use in RotationMatrix. I'll try that to see if it runs any faster.